requirements.txt
twilio
openai
fastapi
uvicorn
pyngrok
python-dotenv
python-multipart